+++ hapi-base/src/main/java/ca/uhn/hl7v2/util/idgenerator/FileBasedHiLoGenerator.java
@@ -40,7 +40,7 @@ public class FileBasedHiLoGenerator extends DelegatingHiLoGenerator {
 		this(DEFAULT_MAXLO);
 	}
 
-	public FileBasedHiLoGenerator(int maxLo) {
+	public FileBasedHiLoGenerator(long maxLo) {
 		super(new FileBasedGenerator(maxLo));
 	}
